[FIX] base_multi_image: Adhere to image delete bypass * Add catch in owner unlink to allow for image delete bypass via context

+++ base_multi_image/models/owner.py
@@ -90,9 +90,12 @@ class Owner(models.AbstractModel):
 
     @api.multi
     def unlink(self):
-        """Mimic `ondelete="cascade"` for multi images."""
+        """Mimic `ondelete="cascade"` for multi images.
+
+        Will be skipped if ``env.context['bypass_image_removal']`` == True
+        """
         images = self.mapped("image_ids")
         result = super(Owner, self).unlink()
-        if result:
+        if result and not self.env.context.get('bypass_image_removal'):
             images.unlink()
         return result
